Remove deb, rpm and tar in favour for the snap.
Motivation:

First of, our packages where not that well maintained and it's a lot of
work to test and validate library capability cross different Debian
based versions for the deb.

I have no idea if even the tar-ball or the rpm even works.

The snap is available with a simple "snap install konstructs-client".
snap is pre-installed on all current Ubuntu installations and it's super
easy to install it on Debian, Fedora, Arch, ...

I know that there are a few people out there who prefer flatpacks and it
think it's better time spent to make a flatpack instead of maintaining
this old code.

Flatpack and snaps are the way forward!